requirements.txt
aiobotocore==2.6.0
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beanie==1.21.0
botocore==1.31.17
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
lazy-model==0.1.0b0
motor==3.2.0
multidict==6.0.4
passlib==1.7.4
phonenumbers==8.13.19
pydantic==2.1.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.4.0
PyJWT==2.8.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
wrapt==1.15.0
yarl==1.9.2
